Add: new option: "evalOnlyFloat"
This option is to forbid using non-float type
values/variables/functions(return value) in inputted expressions.

This option is strongly recommended to be enabled
when "evalNumberAsFloat" option is enabled,
to prevent overlooking "contermination" of integer values,
which can cause of unexpected integer divisions.

The "evalNumberAsFloat" option is enabled by default,
so this option is also enabled by default.

This option does not affect to the interpretation of
the code in the library script files.