Remove: CodeLoader and PluginLoader
On old versions, library scripts and plug-ins were loaded from files
on the application-side, and were passed to the script engine.

On the other hand, on recent versions,
the script engine loads library scripts and plug-ins from files directly.

Therefore, loaders of them are no longer necessary on the application-side,
so they have been removed by this commit.