Change: the content of the example library script
We had added "Example.vnano" as an example of a practical script
at the previous commit, so we simplified the content of
"/lib/ExampleLibrary.vnano" for focusing on the demonstration of
how to define a variable and a function.

Also, names of old variables/functions in "/lib/ExampleLibrary.vnano"
had been too simple, so they could have conflicted with
variable/functions in other scripts.
Therefore they have changed by this commit.