Fix: behaviour when a new calculation is requested when an previous calculation is running
On the correct behaviour, if a new calculation is requested
when an calculation requested previously is still running,
the RINPn should decline it with displaying a message
"The previous calculation has not finished yet".

However, on recent versions,
the RINPn had not declined such additional calculation request,
and it had executed after when the running calculation completed.

The above behaviour is incorrect. It have been caused by
"synchronized" modifier of the method performing calculation.
( The calculation procedure on the GUI mode had originally
  implemented in AsyncronousCalculationRunner class as an
  non-"synchronized" method, but some part of it was moved
  into CalculatorModel class as an "synchronized" method,
  by the refactoring taken at an old version.)

This incorrect behaviour has been fixed by this commit.
Also, the processing flow to invoke the calculation method
on the GUI mode has been simplified a little.