Change: rename "Setting.vnano" to "Settings.txt"
Perhaps, it is not easy for many users
to open the file having the extension ".vnano"
which is not associated with text editors
by default.

On the other hand, probably, almost all of users
can open the file having the extension ".txt"
very easily by double-clicking the file.

We want that users open the setting file
and grasp what are customizable,
so we have changed the extension of
the setting file from ".vnano" to ".txt".