Add: source code of the scripting engine of Vnano
The source code of the scripting engine is maintained
on the other repository, and anyone can get it anytime.
So we had considered that
there is no need to bandle it in this repository.
However, we have changed our mind,
and have bundled it in this repository by this commit.

The main reason is:
Each version of this app requires the corresponding
version of the scripting engine.

The secondary reason is:
We want to make the building steps simple and easy,
as possible as.