Change: ViewImpl: make all fields public, and remove getters/setters
When we had designed the internal structure of this app,
we had separated the "view" layer to the implementation (ViewImpl)
and the interface (ViewInterface).
Until this commit, the "presenter" layer has accessed to
the UI components through the getters/setters
defined by the ViewInterface.

However, the current structure is too complex,
for this simple app.
We want to simplify the implementation.

Hence, we have changed the all UI elements in the ViewImpl "public".
Now the presenter layers accesses to them directly,
not through the ViewInterface.