Fix: key panel had not considered the caret
When we click any key,
the corresponding number/word should be inputted
at just back of the caret, in "INPUT" text field.

However, it had been appended to
the end of the content of the text field,
with ignoring the caret position.

The above has been fixed by this commit.